Create pypi wheel for opm-python
This version tries to use the generated pybind11 modules opmcommon_python.so
and simulators.so from the host system, but auditwheel in the
manylinux docker container will not be able to repair those if they
were compiled with a too recent toolchain on the host.

I think we need to build both pybind11 modules inside the manylinux
docker container also.